refactor(mqtt): 重构 MQTT 客户端实现以支持多平台
- 将 MQTT 相关代码从 utils/mqtt.dart 移至 utils/mqtt/ 目录
- 创建平台特定的 MQTT 客户端工厂实现
- 使用 createMqttClient 函数统一处理 Web 和原生平台的客户端创建
- 移除 kIsWeb 条件判断逻辑，改用平台检测导入
- 更新生产任务视图中的导入路径以匹配新的目录结构

+++ lib/bean/http/response/production_tasks_info.dart
@@ -90,8 +90,8 @@ class ProductionTasksSubInfo {
   bool? isClose;
   String? itemImage; //型体图片
   double? finishQtyTotal; //累计完成数
-  int? shouldPackQty; //应装箱数
-  int? packagedQty; //装箱数
+  double? shouldPackQty; //应装箱数
+  double? packagedQty; //装箱数
   bool? existOutBoxBarCode; //外箱标扫码状态
   List<String>? packetWay; //装箱方式
   List<String>? specificRequirements; //客人需求
@@ -272,8 +272,8 @@ class WorkCardSizeInfos {
 class ProductionTasksDetailInfo {
   String? billNo;
   String? clientOrderNumber;
-  int? total;
-  int? hasInstall;
+  double? total;
+  double? hasInstall;
   List<String>? packetWay; //装箱方式
   List<String>? specificRequirements; //客人需求
   List<ProductionTasksDetailItemInfo>? scheduleInfos;
